alt-bot-service/requirements.txt
fastapi~=0.85.0
uvicorn
pydantic~=1.10.2
mysql-connector-python~=8.0.30
certifi==2022.9.24
requests~=2.28.1
numpy~=1.23.3
schedule